apps: move device related operation to a new file
